#767 refactor and moved files
